+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_BrowserAutomotion.cs
@@ -65,7 +65,7 @@ public class Sample_BrowserAutomotion : ProjectsOpenAITestBase
                 new BrowserAutomationToolConnectionParameters(playwrightConnection.Id)
             ));
 
-        PromptAgentDefinition agentDefinition = new(model: modelDeploymentName)
+        DeclarativeAgentDefinition agentDefinition = new(model: modelDeploymentName)
         {
             Instructions = "You are an Agent helping with browser automation tasks.\n" +
             "You can answer questions, provide information, and assist with various tasks\n" +
@@ -129,7 +129,7 @@ public class Sample_BrowserAutomotion : ProjectsOpenAITestBase
                 new BrowserAutomationToolConnectionParameters(playwrightConnection.Id)
             ));
 
-        PromptAgentDefinition agentDefinition = new(model: modelDeploymentName)
+        DeclarativeAgentDefinition agentDefinition = new(model: modelDeploymentName)
         {
             Instructions = "You are an Agent helping with browser automation tasks.\n" +
             "You can answer questions, provide information, and assist with various tasks\n" +