+++ sdk/recoveryservices/Azure.ResourceManager.RecoveryServices/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.2.0-beta.2 (Unreleased)
+## 1.3.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,18 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.2.0 (2026-03-20)
+
+### Features Added
+
+- Upgraded api-version to 2025-08-01.
+- Make `Azure.ResourceManager.RecoveryServices` AOT-compatible.
+
+### Other Changes
+
+- Upgraded dependent `Azure.Core` to 1.51.1.
+- Upgraded dependent `Azure.ResourceManager` to 1.14.0.
 
 ## 1.2.0-beta.1 (2024-10-15)